resolved issues from merge, fixed pax code to work with changed methods in ColumnStoreHelper

+++ src/pax_store/Page.h
@@ -174,14 +174,14 @@ class PaxPage {
     // Note: This function currently only works when T is set to uint64_t. Otherwise, the `ColumStore::Helper::store` function does not
     // work.
     // TODO: Allow for uint32_t values
-    std::tuple<uint64_t *, uint64_t> queryAVX(std::vector<Filters::Filter<T, SIMD::AVX512> *> &filters) {
+    std::tuple<uint32_t *, uint64_t> queryAVX(std::vector<Filters::Filter<T, SIMD::AVX512> *> &filters) {
         // Array of all row indices which have passed all filters that were already applied. At the beginning, all rows all still present
         // because no filter has been applied yet. However, for performance reasons, we do not populate the array with all indicies but
         // rather compute the indices on the fly when the first filter is applied (firstRun=true).
         // Please note that, for performance reasons, we do not recreate the positions array after each iteration. Instead, we overwrite the
         // elements starting from the beginning of each iteration and keep track of all still valid positions using the `positionsCounter`.
         // All values with an index greater than `positionsCounter` are outdated intermediate results.
-        auto positions = new uint64_t[*numberOfRecords];
+        auto positions = new uint32_t[*numberOfRecords];
         bool firstRun = true;
 
         // Count of all rows which have passed all filters.
